requirements.txt
numpy==1.21.5
Pillow==11.1.0
